+++ src/Hyyan/WPI/Product/Meta.php
@@ -10,7 +10,9 @@
 
 namespace Hyyan\WPI\Product;
 
-use Hyyan\WPI\HooksInterface;
+use Hyyan\WPI\HooksInterface,
+    Hyyan\WPI\Utilities,
+    Hyyan\WPI\Admin\Settings;
 
 /**
  * Prodcut Meta
@@ -43,7 +45,7 @@ class Meta
     {
 
         // sync product meta with polylang
-        add_filter('pll_copy_post_metas', array($this, 'defineProductMetaToCopy'));
+        add_filter('pll_copy_post_metas', array(__CLASS__, 'getProductMetaToCopy'));
 
         $currentScreen = get_current_screen();
 
@@ -62,9 +64,6 @@ class Meta
          *
          * if the "new_lang" is defined or if the current page is the "edit"
          * page then porduct meta editing must by disabled
-         *
-         * enqueue the lock-fileds.js script if we should disable porduct
-         * meta editing
          */
 
         if (isset($_GET['post'])) {
@@ -78,7 +77,11 @@ class Meta
 
         // disable fields edit for translation
         if ($disable) {
-            $this->addFieldsLockerScript();
+            add_action(
+                    'admin_print_scripts'
+                    , array($this, 'addFieldsLocker')
+                    , 100
+            );
         }
 
         /* sync selected prodcut type */
@@ -89,59 +92,139 @@ class Meta
      * Define the meta keys that must copyied from orginal product to its
      * translation
      *
-     * @param array $metas array of meta keys
+     * @param array   $metas array of meta keys
+     * @param boolean $flat  false to return meta list with sections (default true)
      *
      * @return array extended meta keys array
      */
-    public function defineProductMetaToCopy(array $metas = array())
+    public static function getProductMetaToCopy(array $metas = array(), $flat = true)
     {
 
-        $default = array(
+        $default = apply_filters(HooksInterface::PRODUCT_META_SYNC_FILTER, array(
             // general
-            'product-type',
-            '_virtual',
-            '_downloadable',
-            '_sku',
-            '_regular_price',
-            '_sale_price',
-            '_sale_price_dates_from',
-            '_sale_price_dates_to',
-            '_downloadable_files',
-            '_download_limit',
-            '_download_expiry',
-            '_download_type',
+            'general' => array(
+                'name' => __('General Metas', 'woo-poly-integration'),
+                'desc' => __('General Metas', 'woo-poly-integration'),
+                'metas' => array(
+                    'product-type',
+                    '_virtual',
+                    '_downloadable',
+                    '_sku',
+                    '_regular_price',
+                    '_sale_price',
+                    '_sale_price_dates_from',
+                    '_sale_price_dates_to',
+                    '_downloadable_files',
+                    '_download_limit',
+                    '_download_expiry',
+                    '_download_type',
+                    'menu_order',
+                    'comment_status',
+                    '_upsell_ids',
+                    '_crosssell_ids',
+                    '_featured',
+                    '_thumbnail_id',
+                    '_price',
+                    '_product_image_gallery',
+                    'total_sales',
+                    '_translation_porduct_type',
+                    '_visibility',
+                )
+            ),
             // stock
-            '_manage_stock',
-            '_stock',
-            '_backorders',
-            '_stock_status',
-            '_sold_individually',
+            'stock' => array(
+                'name' => __('Stock Metas', 'woo-poly-integration'),
+                'desc' => __('Stock Metas', 'woo-poly-integration'),
+                'metas' => array(
+                    '_manage_stock',
+                    '_stock',
+                    '_backorders',
+                    '_stock_status',
+                    '_sold_individually',
+                )
+            ),
             // shipping
-            '_weight',
-            '_length',
-            '_width',
-            '_height',
-            'product_shipping_class',
-            // advanced
-            '_purchase_note',
-            'menu_order',
-            'comment_status',
-            // extra
-            '_upsell_ids',
-            '_crosssell_ids',
-            '_featured',
-            '_thumbnail_id',
-            '_price',
-            '_product_image_gallery',
-            'total_sales',
-            '_translation_porduct_type',
-            '_visibility',
+            'shipping' => array(
+                'name' => __('ShippingClass Metas', 'woo-poly-integration'),
+                'desc' => __('ShippingClass Metas', 'woo-poly-integration'),
+                'metas' => array(
+                    '_weight',
+                    '_length',
+                    '_width',
+                    '_height',
+                    'product_shipping_class',
+                )
+            ),
+            // attributes
+            'Attributes' => array(
+                'name' => __('Attributes Metas', 'woo-poly-integration'),
+                'desc' => __('Attributes Metas', 'woo-poly-integration'),
+                'metas' => array(
+                    '_product_attributes',
+                    '_default_attributes',
+                ),
+            )
+        ));
+
+        if (false === $flat) {
+            return $default;
+        }
+
+        foreach ($default as $ID => $value) {
+            $metas = array_merge($metas, Settings::getOption(
+                    $ID , 'wpi-metas-list' , $value['metas']
+            ));
+        }
+
+        return array_values($metas);
+    }
+
+    /**
+     * Add the Fields Locker script
+     *
+     * The script will disable editing of some porduct metas for product
+     * translation
+     *
+     * @return boolean false if the fields locker feature is disabled
+     */
+    public function addFieldsLocker()
+    {
+
+        if ('off' === Settings::getOption('fields-locker', 'wpi-features', 'on')) {
+            return false;
+        }
+
+        $metas = static::getProductMetaToCopy();
+        $selectors = apply_filters(HooksInterface::FIELDS_LOCKER_SELECTORS_FILTER, array(
+            '.insert',
+            in_array('_product_attributes', $metas) ? '#product_attributes :input' : rand(),
+        ));
+
+        $jsID = 'product-fields-locker';
+        $code = sprintf(
+                'var disabled = %s;'
+                . 'for (var i = 0; i < disabled.length; i++) {'
+                . ' $('
+                . '     %s + ","'
+                . '     + "." + disabled[i] + ","'
+                . '     + "#" +disabled[i] + ","'
+                . '     + "*[name^=\'"+disabled[i]+"\']"'
+                . ' )'
+                . '     .attr("onclick", event.preventDefault())'
+                . '     .css({'
+                . '         opacity: .5,'
+                . '         \'pointer-events\': \'none\','
+                . '         cursor: \'not-allowed\''
+                . '     }'
+                . ' );'
+                . '}'
+                , json_encode($metas)
+                , !empty($selectors) ?
+                        json_encode(implode(',', $selectors)) :
+                        array(rand())
         );
 
-        return array_merge(
-                $metas
-                , apply_filters(HooksInterface::PRODUCT_META_SYNC_FILTER, $default)
-        );
+        Utilities::jsScriptWrapper($jsID, $code);
     }
 
     /**
@@ -171,42 +254,26 @@ class Meta
          * list
          */
         if ($ID && ($type = get_post_meta($ID, '_translation_porduct_type'))) {
+
             add_action('admin_print_scripts', function () use ($type) {
-                printf(
-                        '<script type="text/javascript" id="woo-poly">'
-                        . '// <![CDATA[ %1$s'
+
+                $jsID = 'product-type-sync';
+                $code = sprintf(
+                        '// <![CDATA[ %1$s'
                         . ' addLoadEvent(function () { %1$s'
-                        . '     jQuery("#product-type option").removeAttr("selected");%1$s'
-                        . '     jQuery("#product-type option[value=\"%2$s\"]").attr("selected", "selected");%1$s'
+                        . '  jQuery("#product-type option")'
+                        . '     .removeAttr("selected");%1$s'
+                        . '  jQuery("#product-type option[value=\"%2$s\"]")'
+                        . '         .attr("selected", "selected");%1$s'
                         . '})'
                         . '// ]]>'
-                        . '</script>'
                         , PHP_EOL
                         , $type[0]
                 );
+
+                Utilities::jsScriptWrapper($jsID, $code, false);
             }, 11);
         }
-    }
-
-    /**
-     * Add the Fields Locker script
-     *
-     * The script will disable editing of some porduct metas for product
-     * translation
-     *
-     * @todo Add option to control this part
-     */
-    protected function addFieldsLockerScript()
-    {
-        add_action('admin_enqueue_scripts', function () {
-            wp_enqueue_script(
-                    'hyyan-wpi-fields-locker.js'
-                    , plugins_url('public/js/FieldsLocker.js', Hyyan_WPI_DIR)
-                    , array('jquery')
-                    , \Hyyan\WPI\Plugin::getVersion()
-                    , false
-            );
-        }, 100);
     }
 
 }