+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+###0.21
+
+* Added admin interface to allow user to control plugin features
+* Added link for every attribute to search for its translation in the polylang strings table
+* Added generic fields locker
+* Added POT file for translation
+* Fixed product_type is not synced in 0.20 version
+* General code improvements
+
 ###0.20
 
 * Added the ability to sync total_sales when stock value is changed